Move reproducibility tests into their own file
